Better handling of arguments and junk
`./script ` will try to open ./s1zoom_0 and stitch all captures
dir can be specified positionally or with -d (which takes precedence).

If no origin/center tile is specified with -o --origin it will try to find the middle one.
If no size is specified with -s --size it will use everything.
If you specify an origin with no size, it will generate an image centered on that your origin tile using the same size as it would have. The image will shift toward your origin point, adding black tiles in that direction to fill the void and ignoring whatever might have been in the corner you moved away from. Quirky but not broken behavior. I added a warning and made it not error out in that case by clumsily copying code.